+++ GWToolboxdll/GWToolbox/Modules/InventoryManager.cpp
@@ -750,7 +750,7 @@ void InventoryManager::Draw(IDirect3DDevice9* device) {
 					ImGui::SetTooltip("%s",pi->desc_s.c_str());
 				}
 				ImGui::SameLine(longest_item_name_length + wiki_btn_width);
-				ImGui::PushID(pi->item_id);
+				ImGui::PushID(static_cast<int>(pi->item_id));
 				if (ImGui::Button("Wiki", ImVec2(wiki_btn_width, 0))) {
 					std::wstring wiki_item = std::regex_replace(pi->short_name, std::wregex(L" "), std::wstring(L"_"));
 					wiki_item = std::regex_replace(wiki_item, wsanitiser, std::wstring(L""));
